// schuller app/Back-end/src/model/ScanTicket.ts
export class ScanTicket {
  // decodedString?: string;
  gender?: string;
  first_name?: string;
  last_name?: string;
  fullname?: string;
  company?: string;
  companies?: string[];
  country?: string;
  city?: string;
  cities?: string[];
  zip_code?: string;
  zip_codes?: string[];
  sales_contact_ids?: string[];
  sales_contact_id?: string;
}